Refactor `commands` command
