+++ cli/commands/help.py
@@ -2,4 +2,4 @@ cli_docs = '''
 Hello World
 '''
 
-print(cli_docs)
+#print(cli_docs)
